# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "version": "1"
    },
    "paths": {
        "/api/admin-panel/users/{user_id}": {
            "get": {
                "tags": [
                    "Admin: users"
                ],
                "summary": "Returns datas of current user - > all datas from all resourec (calls,transactions)",
                "description": "Returns datas of current user - > all datas from all resourec (calls,transactions)",
                "operationId": "App\\Http\\Controllers\\Dashboard\\UserController::show",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": [
                                                {
                                                    "id": 63,
                                                    "full_name": null,
                                                    "photo": null,
                                                    "email": null,
                                                    "date_of_birth": "06.09.2021",
                                                    "phone_number": "+3712608379423",
                                                    "gender": null,
                                                    "type": "listener",
                                                    "rating": null,
                                                    "status": "offline",
                                                    "bio": null,
                                                    "email_verified_at": null,
                                                    "device_token_customer": null,
                                                    "device_token_listener": null,
                                                    "remember_token": null,
                                                    "created_at": "2021-08-11T16:33:32.000000Z",
                                                    "updated_at": "2021-08-11T16:34:18.000000Z",
                                                    "balance": 30,
                                                    "profile_url": null,
                                                    "language": null,
                                                    "stripe_id": null,
                                                    "last_activity_date": null,
                                                    "consent_to_agreement": 0,
                                                    "succesfull_calls": 1,
                                                    "total_time": 50,
                                                    "conversations_as_listener": [
                                                        {
                                                            "id": 9,
                                                            "caller_id": 1,
                                                            "listener_id": 63,
                                                            "topic_id": 1,
                                                            "channel": "71iR92aeF8N1osXu",
                                                            "token": "YU8bTTWVuiyqIuHkkymCs2xBQ",
                                                            "started_at": "2021-05-13T05:54:32.000000Z",
                                                            "finished_at": "2021-05-13T05:58:32.000000Z",
                                                            "duration": null,
                                                            "status": "requested",
                                                            "created_at": "2021-07-07T01:31:16.000000Z",
                                                            "updated_at": "2021-07-07T01:31:16.000000Z",
                                                            "check_time": "2021-08-17 16:48:21"
                                                        }
                                                    ],
                                                    "conversations_as_caller": [
                                                        {
                                                            "id": 17,
                                                            "caller_id": 63,
                                                            "listener_id": 64,
                                                            "topic_id": 5,
                                                            "channel": "TWjG2RsdfRAYuZav",
                                                            "token": "7ubgNOymKOygPtC1B9cTt3oix",
                                                            "started_at": "2021-03-27T15:58:23.000000Z",
                                                            "finished_at": null,
                                                            "duration": 50,
                                                            "status": "finished",
                                                            "created_at": "2021-07-07T01:31:16.000000Z",
                                                            "updated_at": "2021-08-19T22:30:02.000000Z",
                                                            "check_time": null
                                                        }
                                                    ],
                                                    "transactions": [
                                                        {
                                                            "id": 30,
                                                            "payment_id": 15,
                                                            "amount": 110,
                                                            "type": "deposit",
                                                            "created_at": "2020-12-24T09:57:38.000000Z",
                                                            "updated_at": "2020-12-24T09:57:38.000000Z",
                                                            "minutes": 11,
                                                            "user_id": null,
                                                            "laravel_through_key": 63
                                                        }
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin: users"
                ],
                "summary": "Change current user type(listener,customer)",
                "description": "Change current user type(listener,customer)",
                "operationId": "App\\Http\\Controllers\\AdminController::editUsersById",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "listener"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Type is changed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin-panel/users/{user_id}/withdraw": {
            "post": {
                "tags": [
                    "Admin: users"
                ],
                "summary": "Withdraw amount of money from users balance with checks. ( AMOUNT IS IN CENTS )",
                "description": "Withdraw amount of money from users balance with checks. ( AMOUNT IS IN CENTS )",
                "operationId": "App\\Http\\Controllers\\AdminController::withdrawFromUser",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": "1234"
                                    },
                                    "minutes": {
                                        "type": "integer",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Withdraw succesfull."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Low balance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Your balance lower,then withdraw amount."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No withdraw type provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Please provide amount or minutes to withdraw."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin-panel/users/tracking/all": {
            "post": {
                "tags": [
                    "Admin: users"
                ],
                "summary": "Track users online time with propertys",
                "description": "Track users online time with propertys",
                "operationId": "App\\Http\\Controllers\\AdminController::tracking",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_time"
                                ],
                                "properties": {
                                    "start_time": {
                                        "type": "date",
                                        "example": "2022-01-01"
                                    },
                                    "end_time": {
                                        "type": "date",
                                        "example": "2022-01-03"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "Name": {
                                                        "type": "string",
                                                        "example": "Vojislav Pavasović"
                                                    },
                                                    "Phone": {
                                                        "type": "string",
                                                        "example": "+37126083794"
                                                    },
                                                    "Time_Online": {
                                                        "type": "integer",
                                                        "example": "4"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "Name": "Vojislav Pavasović",
                                                    "Phone": "+37126083794",
                                                    "Time_Online": "4"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin-panel/users/{user_id}/tracking": {
            "post": {
                "tags": [
                    "Admin: users"
                ],
                "summary": "Show current users Tracks",
                "description": "Show current users Tracks",
                "operationId": "App\\Http\\Controllers\\AdminController::useridTracking",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_time"
                                ],
                                "properties": {
                                    "start_time": {
                                        "type": "date",
                                        "example": "2022-01-01"
                                    },
                                    "end_time": {
                                        "type": "date",
                                        "example": "2022-01-03"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "full_name": {
                                                        "type": "string",
                                                        "example": "test123"
                                                    },
                                                    "phone_number": {
                                                        "type": "string",
                                                        "example": "+37126083794"
                                                    },
                                                    "start_time": {
                                                        "type": "datetime",
                                                        "example": "2022-01-11 15:39:21"
                                                    },
                                                    "end_time": {
                                                        "type": "datetime",
                                                        "example": "2022-01-12 02:06:45"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "full_name": "test123",
                                                    "phone_number": "+37126083794",
                                                    "start_time": "2022-01-11 15:39:21",
                                                    "end_time": "2022-01-12 02:06:45"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}